requirements.txt
pymysql
sqlalchemy
cryptography